Fix time zone issue for schedule
When expanding a schedule to daily occurrences `dtstart` and `dtend` are messed up.

This is caused by changing `start_time` and `end_time` from the `IceCube` schedule cast in UTC with the `dtstart` and `dtend`fields which are are cast in the time zone of the app.

This fix solves the issue and cleans up the code slightly.